+++ src/main/java/com/chatsul/service/LostItemService/LostItemQueryServiceImpl.java
@@ -1,4 +1,4 @@
-package com.chatsul.service;
+package com.chatsul.service.LostItemService;
 
 import com.chatsul.domain.LostItem;
 import com.chatsul.repository.LostItemRepository;
